Refactored step inheritance to make this less confusing.
